Apply MISRA C++-2008 Rule 5-0-11
:Release Notes:
Apply MISRA C++-2008 Rule 5-0-11

:Detailed Notes:
- Modify to typecasting from char to int when using isdigit, isalpha, toupper

:Testing Performed:
Build Successful

:QA Notes:
N/A

:Issues Addressed:
[PLAT-124143] General Support for Auto Maintenance

+++ src/memorymanager/sysinfo/SysInfo.cpp
@@ -39,10 +39,10 @@ long SysInfo::parseSizeToKb(string size)
     long ret = 0;
 
     for (auto &c : size) {
-        if (isdigit(c))
+        if (isdigit(static_cast<int>(c)))
             value.push_back(c);
-        if (isalpha(c)) {
-            unit.push_back(toupper(c));
+        if (isalpha(static_cast<int>(c))) {
+            unit.push_back(toupper(static_cast<int>(c)));
         }
     }
